Minor tweaks
I added a "device" variable so training/running the model can be done on CPU, NVIDIA GPUs, Apple Silicon and TPU support can also be added easily thanks to PyTorch. Noise was added to the MCTS' search function and temperature was added to self-play function both to encourage exploration alongside with weight decay.

CS. A.